chore: fix enqueueing to wrong next step
